LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-after-relayout/multiple-aligned-targets/prefer-focused-nested-containers-expected.txt

LPH (outer)RPH (outer)

FAIL Snap container prefers focused nested snap target. assert_equals: scroller followed rightcontainer in y axis after layout change expected 600 but got 500
